Set default values for "unset" gconf keys
If a gconf key doesn't exist, doesn't have a schema or is just unset set
the value to the paramspec default. This fixes the picture being black if
no schema installed (brightness and contrast set to 0).
The best thing would be to ask gstreamer for the default as we do in the
preferences dialog but at the moment CheeseGconf is still unaware of
CheeseWebcam.
The method used (check gconf_client_get to be non NULL) is a bit tricky
but that's the only way to get information about unset keys.
Yes, gconf sucks.