window: remove initial black bars (at least for 4:3 webcams)
Set a size request on the drawing area, get the requisition computed on
the main window and use it as default size. Then the first size request
can be removed and the window will keep the size we requested.
Handy trick to set default size of the drawing area without limiting its
minimum size.