webcam: fix broken xoverlay management
For some reason this went unnoticed for a long time. We were calling
set_x_overlay right after setting PLAYING state (with no warranty that
the state change succeeded nor that an overlay was really there). The
only right way to do this is to wait for prepare-xwindow-id message. It
worked until now just because we've been extremely lucky.