thumbview: let iconview pick up its optimal size
Remove hard coded size constraints for the icon view so that it picks up
the best size for the current children (i.e. centered items, even with
no selection, with every thumbnail size).
It's almost an hack but seems to work pretty well here. The alternative
is to keep the current size request but it's quite ugly.
Remove also extra spacing between and around thumbnails.