Fix possible crasher in CheeseCamera when capturing pixbufs
Craziness. We were creating a pixbuf from a GstBuffer, but the
data in the GstBuffer is from GStreamer, and might change at the
end of the function.

We need to duplicate the buffer if we want the pixbuf to stay
static, otherwise funny effects like a cropped copy of the
pixbuf reflecting the live video.