Restore temporarily disabled dbus code
We should really remove dbus stuff. We agreed to use libcheese for
interoperation with the desktop so we're keeping it just for the single
instance stuff. Probably gunique would be better suited for it.
The best would be to make CheeseCameraDevice aware of device locking so
we wouldn't need to be single instance anymore.