remove the effect files from cheese
from now on, effects are brought to the users home by the
package gnome-video-effects (http://git.gnome.org/browse/gnome-video-effects).
this way, many applications, like Pitivi or Empathy are able
to use the same effects as we do. however, we will not
depend on it, as it is an optional package and not needed to
run.

cheese looks for a directory named gnome-video-effects in
one of the system share directories proposed by XDG (mostly
/usr/share/ or /usr/local/share/).

it also looks if the user has a directory named
~/.gnome-video-effects. we probably should change that and
use the XDG standard for user shared files, which is ~/.local/share/