thumb-view: handle photo/video moved to monitored directories
Camerabin2 creates a temporary file, writes data to it and in the end
it moves it to the destination path.

Without G_FILE_MONITOR_SEND_MOVED to g_file_monitor_directory, moves
would be sent as DELETE + CREATE.

Unfortunately, we would also get CREATE events when the temporary file
was created. We cannot listen for CREATE events as temporary files
would be CREATEd and then CHANGEd and the thumb-view creation code
would work with bad data.

By using G_FILE_MONITOR_SEND_MOVED we get a single event when the file
is moved. As the file monitor ignores files with unknown extensions
(as is the case for the temporary files created by camerabin2) we
could only just append the new file. But because the user might also
move a picture/video manually from the Pictures/Webcam or
Videos/Webcam directories, removing the old file from the thumb-view
keeps it up-to-date with the state of the file system.

If camerabin2 switches away from the create-temp+move approach, to the
camerabin one (create the destination file directly), we're prepared
to handle that code with G_FILE_MONITOR_EVENT_CHANGES_DONE_HINT.