Do not overwrite camerabin's default flags
The comment above this code said: "Set flags to enable conversions", but
camerabin has conversion enabled in the necessary places by default, all
the code does is add an extra, unneeded ffmpegcsp element at the
beginning of the pipeline.

This results in a huge number of reported caps on pads further down the
pipeline, which when intersected with input caps of later ffmpegcsp
elements leads to an explosion of possible combinations and
gst_caps_intersect_full starts consuming the CPU for 100% for seconds
(various Fedora users have reported startup delays of upto a minute).

On my test system, with a Logitech Webcam 9000 Pro, the time to create
the camerabin pipeline (not start, not configure, just create!) drops
from 7 seconds to 0.7 seconds by elimenating the unnecessary ffmpegcsp
element at the beginning of the pipe.

The only reason the ffmpegcsp element this patch removes could be useful
would be for cameras producing only JPEG data, but since Cheese always
uses v4l2src, and that should always be compiled with libv4l2 support
(otherwise a lot of camera specific video formats will not be
understood), libv4l2 will take care of JPEG decompression, so there
really is no reason for having this extra element, and thus no reason to
override the default camerabin flags.

Fixes bug 677731.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>