Do not add 0-sized files to the thumb view
Sometimes, ie when Cheese crashes when starting to record, something which
we need to fix, 0 bytes large files are created under ~/Videos/Webcam.

totem-video-thumbnailer does not like these, getting stuck for around a
minute on them, I have filed bug 677734 for this.

But even with this bug fixed, trying to add 0 byte sized files to the
thumb view makes no sense, and in the mean time it also neatly works
around this totem-video-thumbnailer bug.

Fixes bug 677735.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>