Ensure width is a multiple of 8, and height of 2
We ask GStreamer to use an YUV format in the pipeline, by setting the
camerabin's "filter-caps" to "video/x-raw-yuv", and GStreamer demands of
YUV format frames that their width is a multiple of 8, and their height
a multiple of 2.

Previous to this patch my bttv tvcard lists the following resolutions in
Cheese:
924x576 (*)
920x576
640x480
462x288 (*)
460x288 (*)
320x240
231x144 (*)
230x144 (*)
160x120

Where all the ones marked with an asterisk do not work.

After this patch the list is:
920x576
640x480
456x288
320x240
224x144
160x120

And all work.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>