Fix no-video-after-switching-resolution problem
There is a bug in wrappercamerabinsrc which causes it to lose its
video-source setting after the pipeline has started, so on a
stop/re-start, as when changing resolution, its video-source has become
NULL, and there is no longer video.

This patch works around this by moving the setting of the video-source
property to cheese_camera_play(), so that it gets (re)set each time
before starting the pipeline.

I've also written a patch fixing the underlying cause, but since the
workaround is simple, and has no adverse effects when the underlying
issue is fixed, it seems a good idea to have this workaround in Cheese,
see here for the gst-plugins-bad fix:

https://bugzilla.gnome.org/show_bug.cgi?id=701915

Signed-off-by: Hans de Goede <hdegoede@redhat.com>