Only add videoconvert elements where necessary
The "no effect" effect is our default value, and thus worth optimizing a
bit. Clearly in the "no effect" effect case adding a videoconvert
element both before and after the element is not needed.

Note we also don't add the videoconvert elements when creating the
initial pipeline, so this also keeps the way the pipeline looks
initially and when "no-effect" is selected consistent.

When starting Cheese with "no-effect" selected, this shaves off another
130 ms of the initial pipeline creation time.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>