Downscale image for effects-preview pipeline
Having the whole effects-preview bin deal with ie 1280x800 images is not
very useful, especially since even when fullscreen on a full-HD monitor,
the preview images are smaller then 640xXXX. This useless high-res
processing for 9 preview images in parallel brings my 2nd gen
core i5 @ 3.1 GHz to its knees, resulting in a non fluid preview panel.

Also after clicking through all effect preview pages, so that all effect
preview textures are connected, cheese will use 1 GB of *resident* RAM
with the example 1280x800 capture resolution.

This patch improves this by downscaling the images from the video-source
to 640xXXX where XXX is determined by the original resolution
aspect-ratio.

After this patch the effects preview framerate is much smoother, and the
latency is noticably less. As a bonus the maximal resident size of
Cheese in this example is reduced to 350 MB.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>