Fix being paused when effect valves are closed
The move to autocluttersink in 3.9 causes a regression compared to 3.8,
where if an effect's control valve is closed the pipeline will never
transition from paused to playing since the cluttersink is stuck in
paused.

This regression can be triggered the following way:
1) Go the effect selector
2) Go to the next page in the effect selector, so that the now connected
   effects from the first page get their control valve set to closed
3) Change resolution or device, so that the pipeline gets stopped and
   started again

After 3) the pipeline will be stuck in pause, as the cluttersinks of the
effect previews are stuck in pause.

The cause for this is that with the move to autocluttersink we are no
longer setting the sink's "async" property, see commit 424afdaa 'Use
"autocluttersink" instead of "cluttersink"'.

Setting the "async" property is no longer (easily) doable, since
autocluttersink is a gstbin derived object which has the actual sink
embedded.

Instead this patch fixes the stuck in pause problem by locking the
autocluttersinks of the effect previews into the playing state.

Note that no longer setting of the async property on the sink is also
the root cause of the issue fixed by my patch titled: "Set all valves
open on play". The opening of the main valve is still necessary, since
locking the viewfinder sink's state to PLAYING does not seem a good
idea, the opening of the effects-valve on play OTOH is no longer
necessary.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>