Remove help translations with incompatible license
Since the license of the help was changed to CC-BY-SA 3.0 Unported, the
translations must have license that is compatible. Some licenses were
declared to be the same as the "Cheese package", which would be the GNU
GPL 2 or later, which is incompatible with CC-BY-SA 3.0 Unported.

Remove the problematic translations, as they cannot be distributed with
Cheese unless the license header is updated by the contributors. Fixes
bug 705568/