Correctly select the given camera device
Due to a bug in the VAPI file the code would pass a string to the camera
setup routine instead of a CameraDevice object. As the setup routine
already has code to automatically select an appropriate device based on
the name passing an argument is not even required.

Note that the --device parameter might be somewhat confusing as it
refers to the devices name and not to the /dev/videoX device file.

https://bugzilla.gnome.org/show_bug.cgi?id=777047