build: Remove obsolete X-GNOME-* keys
Cheese's desktop file contains obsolete `X-GNOME-*` keys. Due to
this these keys has been removed.

The desktop file was also used as an input file to replace
`PACKAGE_VERSION` tag, used by `X-GNOME-Bugzilla-Version`. However,
once this key has been removed there is no tag to be replaced so
the input file has been renamed from `.in.in` to `.in` so only the
is processed once to merge translations.